+++ Botcoin/Models/OrderModel.cs
@@ -14,9 +14,9 @@ namespace Botcoin.Models
         [MaxLength(5)]
         public string Coin { get; set; }
 
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
 
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
         public DateTime DateRegistered { get; set; }
     }
